Renamed ReduxViewModel to ReduxViewModelGenerator
